requirements.txt
numpy==1.14.2
scipy==1.0.1
scikit-learn==0.19.1
scikit-surprise==1.0.5
